+++ src/Bio/Uniprot/Parser.hs
@@ -2,13 +2,19 @@
 {-# LANGUAGE RecordWildCards   #-}
 module Bio.Uniprot.Parser where
 
+import           Prelude              hiding (null)
+
 import           Bio.Uniprot.Type
 import           Control.Applicative  ((<|>))
+import           Control.Monad        (unless, when)
 import           Data.Attoparsec.Text
 import           Data.Bifunctor       (second)
 import           Data.Functor         (($>))
-import           Data.Text            (Text, pack, unpack)
+import           Data.Text            (Text, append, init, null, pack, unpack,
+                                       unwords)
 import           Data.Void
+
+import           Debug.Trace          (trace)
 
 parseID :: Parser ID
 parseID = do
@@ -25,16 +31,17 @@ parseID = do
 
 parseAC :: Parser AC
 parseAC = do
-    initAC <- concat <$> many' (parseOneAC <* endOfLine)
+    parseStartAC
+    initAC <- concat <$> many' (parseOneAC <* endOfLine <* parseStartAC)
     lastAC <- parseOneAC
     let accessionNumbers = initAC ++ lastAC
     pure AC{..}
   where
+    parseStartAC :: Parser ()
+    parseStartAC = string "AC" >> count 3 space >> pure ()
+
     parseOneAC :: Parser [Text]
-    parseOneAC = do
-        string "AC"
-        many1 space
-        many1 (parseTextToken <* char ';' <* many' space)
+    parseOneAC = many1 (parseTextToken <* char ';' <* option ' ' (satisfy isHorizontalSpace))
 
 parseDT :: Parser DT
 parseDT = do
@@ -45,7 +52,7 @@ parseDT = do
   where
     parseOneDT :: Text -> Parser (Text, Text)
     parseOneDT txt = do
-        string "DE"
+        string "DT"
         many1 space
         day <- pack <$> many1 (satisfy $ inClass "A-Z0-9-")
         char ','
@@ -55,8 +62,28 @@ parseDT = do
         char '.'
         pure (day, x)
 
-parseDE :: Parser ()
-parseDE = parseAndSkip "DE"
+parseDE :: Parser DE
+parseDE = do
+    recName  <- optional $ parseNameDE "RecName"
+    altNames <- many' (endOfLine *> parseAltDE)
+    subNames <- many' (endOfLine *> parseNameDE "SubName")
+    includes <- pure []
+    contains <- pure []
+    flags <- fmap (read . unpack) <$> optional (endOfLine *> parseDELine (Just "Flags") "")
+    pure DE{..}
+
+parseSQ :: Parser SQ
+parseSQ = do
+    string "SQ" >> count 3 space >> string "SEQUENCE"
+    seqLength <- many1 space *> decimal <* space <* string "AA;"
+    molWeight <- many1 space *> decimal <* space <* string "MW;"
+    crc64 <- pack <$> (many1 space *> many1 (satisfy $ inClass "A-F0-9") <* space <* string "CRC64;")
+    endOfLine
+    sequence <- pack . concat <$> many1 (skipSpace *> parseAminoAcids)
+    pure SQ{..}
+
+parseEnd :: Parser ()
+parseEnd = string "//" >> pure ()
 
 -- Helpers
 
@@ -70,6 +97,41 @@ parseAndSkip txt = do
         string txt
         many1 anyChar
         pure ()
-  
+
+parseAminoAcids :: Parser String
+parseAminoAcids = many1 (satisfy $ inClass "ACDEFGHIKLMNPQRSTVWY")
+
 parseTextToken :: Parser Text
 parseTextToken = pack <$> many1 (satisfy $ inClass "A-Z0-9_")
+
+optional :: Parser a -> Parser (Maybe a)
+optional par = option Nothing (Just <$> par)
+
+-- DE helper parsers
+
+parseNameDE :: Text -> Parser Name
+parseNameDE nameType = do
+    fullName <- parseDELine (Just nameType) "Full"
+    shortName <- many' $ endOfLine *> parseDELine Nothing "Short"
+    ecNumber <- many' $ endOfLine *> parseDELine Nothing "EC"
+    pure Name{..}
+
+parseAltDE :: Parser AltName
+parseAltDE =
+  (Simple <$> parseNameDE "AltName") <|>
+  (Allergen <$> parseDELine (Just "AltName") "Allergen") <|>
+  (Biotech <$> parseDELine (Just "AltName") "Biotech") <|>
+  (CDAntigen <$> parseDELine (Just "AltName") "CD_antigen") <|>
+  (INN <$> parseDELine (Just "AltName") "INN")
+
+parseDELine :: Maybe Text -> Text -> Parser Text
+parseDELine name tpe = do
+    string "DE   "
+    string $ maybe "         " (`append` ": ") name
+    unless (null tpe) $ do
+        string tpe
+        string "="
+        pure ()
+    res <- pack <$> many1 (notChar ';')
+    char ';'
+    pure res