+++ ChangeLog.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 * All Uniprot-KB datalines described
+* Parser for ID, AC, DT, DE (w/o Includes and Contains) and SQ lines